backend/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
JPype1==1.4.1
optapy==8.31.1b0
packaging==23.0
psycopg2-binary==2.9.6
pydantic==1.10.5
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
